Here are the actual changes
Try getting rid of the while loop in the Kights Tour App and running a 4x4 board starting at 1. It should iterate all the way through it. 
If you do the same thing without commenting out the while loop, it acts as an infinite loop because the position or errorstack isn't being reset. havn't quite figured it out.